BOOTSTRAP/FARRAGO/FENNEL:  xmlbeans is actually required for blackhawk; don't try to copy postgresql driver if it isn't there; bump up JVM heap to 256M for server; include initBuild.sh in fennel dist
[git-p4: depot-paths = "//open/dev/": change = 8610]

+++ farrago/distBuild.sh
@@ -147,7 +147,9 @@ cp jgrapht/license-LGPL.txt $LIB_DIR/vjdbc.license.txt
 cp jgrapht/license-LGPL.txt $LIB_DIR/RmiJdbc.license.txt
 cp hsqldb/doc/hypersonic_lic.txt $LIB_DIR/hsqldb.license.txt
 cp hsqldb/lib/hsqldb.jar $LIB_DIR
-cp postgresql-8.1-406.jdbc2.jar $LIB_DIR
+if [ -e postgresql-8.1-406.jdbc2.jar ]; then
+    cp postgresql-8.1-406.jdbc2.jar $LIB_DIR
+fi
 cp commons-transaction-1.1.jar $LIB_DIR
 cp vjdbc/lib/vjdbc.jar $LIB_DIR
 cp vjdbc/lib/vjdbc_server.jar $LIB_DIR
